refactor(chat): convert react-free chat hooks into plain modules
