+++ src/components/Chat/__tests__/Chat.recentMessages.test.tsx
@@ -322,8 +322,8 @@ jest.mock('../hooks/useChatScroll', () => ({
   }),
 }));
 
-jest.mock('../hooks/useChatSevenTvCallbacks', () => ({
-  useChatSevenTvCallbacks: () => ({}),
+jest.mock('../util/createSevenTvCallbacks', () => ({
+  createSevenTvCallbacks: () => ({}),
 }));
 
 jest.mock('../hooks/useEmoteReprocessing', () => ({
