+++ src/components/Chat/hooks/useChatCosmetics.ts
@@ -1,34 +1,17 @@
 import { useEffect } from 'react';
 
-import { useLazyRef } from '@app/hooks/useLazyRef';
 import { sevenTvService } from '@app/services/seventv-service';
-import {
-  fetchAndCacheUserCosmetics,
-  getUserBadge,
-  getUserBadgeId,
-  getUserPaintId,
-  requestUserCosmeticsViaPresence,
-} from '@app/store/chat/actions/cosmetics';
+import { fetchAndCacheUserCosmetics } from '@app/store/chat/actions/cosmetics';
+import { hasRenderableCosmetics } from '@app/store/chat/actions/userCosmeticsFetch';
 import { logger } from '@app/utils/logger';
 
-import { boundedSetAdd } from '../util/hydrateVisibleSevenTvAssets/boundedSetAdd';
-
-const MAX_FETCHED_COSMETICS_USERS = 500;
-
-function getRenderableBadgeUrl(twitchUserId: string): string | undefined {
-  const badgeId = getUserBadgeId(twitchUserId);
-  return badgeId ? getUserBadge(twitchUserId)?.url?.trim() : undefined;
-}
-
-function hasRenderableCosmetics(twitchUserId: string): boolean {
-  return Boolean(
-    getUserPaintId(twitchUserId) || getRenderableBadgeUrl(twitchUserId),
-  );
-}
-
+/**
+ * Bootstraps the signed-in user's own 7TV cosmetics once per chat session.
+ * Chatter cosmetics fetching lives in
+ * `store/chat/actions/userCosmeticsFetch` - import it directly.
+ */
 export function useChatCosmetics(options: { userId?: string | null } = {}) {
   const { userId } = options;
-  const fetchedCosmeticsUsersRef = useLazyRef(() => new Set<string>());
 
   useEffect(() => {
     if (!userId || hasRenderableCosmetics(userId)) {
@@ -55,50 +38,4 @@ export function useChatCosmetics(options: { userId?: string | null } = {}) {
       cancelled = true;
     };
   }, [userId]);
-
-  const fetchUserCosmetics = async (
-    twitchUserId: string,
-    fetchOptions: {
-      retryMissingBadge?: boolean;
-    } = {},
-  ) => {
-    const existingBadgeId = getUserBadgeId(twitchUserId);
-    const renderableBadge = getRenderableBadgeUrl(twitchUserId);
-
-    if (
-      fetchedCosmeticsUsersRef.current.has(twitchUserId) &&
-      (!fetchOptions.retryMissingBadge || renderableBadge || !existingBadgeId)
-    ) {
-      return;
-    }
-
-    if (hasRenderableCosmetics(twitchUserId)) {
-      boundedSetAdd(
-        fetchedCosmeticsUsersRef.current,
-        twitchUserId,
-        MAX_FETCHED_COSMETICS_USERS,
-      );
-      return;
-    }
-
-    boundedSetAdd(
-      fetchedCosmeticsUsersRef.current,
-      twitchUserId,
-      MAX_FETCHED_COSMETICS_USERS,
-    );
-
-    try {
-      await requestUserCosmeticsViaPresence(twitchUserId);
-    } catch (error) {
-      logger.stv.debug(
-        `Failed to fetch cosmetics for user ${twitchUserId}:`,
-        error instanceof Error ? error.message : error,
-      );
-    }
-  };
-
-  return {
-    fetchedCosmeticsUsersRef,
-    fetchUserCosmetics,
-  };
 }
