+++ src/components/Chat/hooks/useChatSession.ts
@@ -91,7 +91,7 @@ export function useChatSession({
     });
   }, [channelName, user?.id, user?.login]);
 
-  const { fetchedCosmeticsUsersRef, fetchUserCosmetics } = useChatCosmetics({
+  useChatCosmetics({
     userId: user?.id,
   });
 
@@ -206,7 +206,6 @@ export function useChatSession({
     handleViewableMessagesChange,
   } = useChatMessageProcessing({
     channelId,
-    fetchUserCosmetics,
     handleNewMessage,
     isAtBottomRef,
     maintainBottomAfterContentChange,
@@ -290,7 +289,6 @@ export function useChatSession({
     cleanupScroll,
     cleanupMessages,
     cancelEmoteLoad,
-    fetchedCosmeticsUsersRef,
     isMountedRef: isChatMountedRef,
     processedMessageIdsRef,
   });
@@ -332,7 +330,6 @@ export function useChatSession({
   return {
     connected,
     emoteLoadStatus,
-    fetchUserCosmetics,
     forceFlush,
     getUserState,
     handleViewableMessagesChange,
