+++ src/components/Chat/util/__tests__/createSevenTvCallbacks.test.ts
@@ -1,5 +1,6 @@
-import { act, renderHook } from '@testing-library/react-native';
+import { act } from '@testing-library/react-native';
 
+import { createSevenTvEmote } from '@app/components/Chat/hooks/__tests__/__fixtures__/useChat.fixture';
 import { countMetric } from '@app/lib/sentry';
 import {
   addBadge,
@@ -18,8 +19,7 @@ import type { SanitisedBadgeSet } from '@app/types/twitch/badge';
 import { generateStvEmoteNotice } from '@app/utils/emote/stv/generateSevenTvEmoteNotice';
 import { normalizeSevenTvPaint } from '@app/utils/seventv/cosmetics/normalizeSevenTvPaint';
 
-import { useChatSevenTvCallbacks } from '../useChatSevenTvCallbacks';
-import { createSevenTvEmote } from './__fixtures__/useChat.fixture';
+import { createSevenTvCallbacks } from '../createSevenTvCallbacks';
 import {
   createBadgeChangeEntry,
   createBadgeCosmeticCreateData,
@@ -34,7 +34,7 @@ import {
   createPaintCosmeticUpdateData,
   createPaintInput,
   createPaintPushedEntry,
-} from './__fixtures__/useChatSevenTvCallbacks.fixture';
+} from './__fixtures__/createSevenTvCallbacks.fixture';
 
 jest.mock('@app/store/chat/actions/cosmetics', () => ({
   addBadge: jest.fn(),
@@ -103,37 +103,33 @@ const defaultProps = {
   onEmoteNotice: mockOnEmoteNotice,
 };
 
-describe('useChatSevenTvCallbacks', () => {
+describe('createSevenTvCallbacks', () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
   describe('return shape', () => {
     test('returns all callbacks and channel/set ids', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
-      expect(result.current.onEmoteUpdate).toBeDefined();
-      expect(result.current.onCosmeticCreate).toBeDefined();
-      expect(result.current.onEntitlementCreate).toBeDefined();
-      expect(result.current.onCosmeticUpdate).toBeDefined();
-      expect(result.current.onCosmeticDelete).toBeDefined();
-      expect(result.current.onEntitlementUpdate).toBeDefined();
-      expect(result.current.onEntitlementDelete).toBeDefined();
-      expect(result.current.twitchChannelId).toBe('twitch-123');
-      expect(result.current.sevenTvEmoteSetId).toBe('emote-set-1');
+      expect(result.onEmoteUpdate).toBeDefined();
+      expect(result.onCosmeticCreate).toBeDefined();
+      expect(result.onEntitlementCreate).toBeDefined();
+      expect(result.onCosmeticUpdate).toBeDefined();
+      expect(result.onCosmeticDelete).toBeDefined();
+      expect(result.onEntitlementUpdate).toBeDefined();
+      expect(result.onEntitlementDelete).toBeDefined();
+      expect(result.twitchChannelId).toBe('twitch-123');
+      expect(result.sevenTvEmoteSetId).toBe('emote-set-1');
     });
   });
 
   describe('onEmoteUpdate', () => {
     test('calls updateSevenTvEmotes with channelId, added, removed', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onEmoteUpdate({
+        result.onEmoteUpdate({
           channelId: 'c1',
           added: [
             createSevenTvEmote({ id: 'e1', name: 'e1', original_name: 'e1' }),
@@ -152,9 +148,7 @@ describe('useChatSevenTvCallbacks', () => {
     });
 
     test('emits notice messages for added and removed emotes', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       const added = [
         createSevenTvEmote({ id: 'e1', name: 'Added', original_name: 'Added' }),
@@ -168,7 +162,7 @@ describe('useChatSevenTvCallbacks', () => {
       ];
 
       act(() => {
-        result.current.onEmoteUpdate({
+        result.onEmoteUpdate({
           channelId: 'c1',
           added,
           removed,
@@ -189,9 +183,7 @@ describe('useChatSevenTvCallbacks', () => {
     });
 
     test('suppresses visible notices for nnys emote changes', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       const added = [
         createSevenTvEmote({
@@ -209,7 +201,7 @@ describe('useChatSevenTvCallbacks', () => {
       ];
 
       act(() => {
-        result.current.onEmoteUpdate({
+        result.onEmoteUpdate({
           channelId: 'c1',
           added,
           removed,
@@ -228,22 +220,18 @@ describe('useChatSevenTvCallbacks', () => {
 
   describe('onCosmeticCreate', () => {
     test('no-ops when cosmetic.object is missing', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
         // @ts-expect-error -- exercising runtime guard when cosmetic.object is absent
-        result.current.onCosmeticCreate({ kind: 'BADGE', cosmetic: {} });
+        result.onCosmeticCreate({ kind: 'BADGE', cosmetic: {} });
       });
 
       expect(applyCosmeticCreateEvent).not.toHaveBeenCalled();
     });
 
     test('delegates badge creates to applyCosmeticCreateEvent', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
       const data = createBadgeCosmeticCreateData(
         createBadgeData({
           id: 'badge-id',
@@ -253,7 +241,7 @@ describe('useChatSevenTvCallbacks', () => {
       );
 
       act(() => {
-        result.current.onCosmeticCreate(data);
+        result.onCosmeticCreate(data);
       });
 
       expect(mockApplyCosmeticCreateEvent.mock.calls).toEqual([
@@ -262,9 +250,7 @@ describe('useChatSevenTvCallbacks', () => {
     });
 
     test('delegates paint creates to applyCosmeticCreateEvent', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
       const data = createPaintCosmeticCreateData(
         createPaintInput({
           id: 'paint-id',
@@ -274,7 +260,7 @@ describe('useChatSevenTvCallbacks', () => {
       );
 
       act(() => {
-        result.current.onCosmeticCreate(data);
+        result.onCosmeticCreate(data);
       });
 
       expect(mockApplyCosmeticCreateEvent.mock.calls).toEqual([
@@ -285,9 +271,7 @@ describe('useChatSevenTvCallbacks', () => {
 
   describe('onEntitlementCreate', () => {
     test('delegates to applyEntitlementCreateEvent', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
       const data = {
         entitlement: {
           id: 'entitlement-1',
@@ -314,7 +298,7 @@ describe('useChatSevenTvCallbacks', () => {
       };
 
       act(() => {
-        result.current.onEntitlementCreate(data);
+        result.onEntitlementCreate(data);
       });
 
       expect(mockApplyEntitlementCreateEvent.mock.calls).toEqual([[data]]);
@@ -323,12 +307,10 @@ describe('useChatSevenTvCallbacks', () => {
 
   describe('onCosmeticUpdate', () => {
     test('records a Sentry count metric for paint updates', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onCosmeticUpdate(
+        result.onCosmeticUpdate(
           createPaintCosmeticUpdateData({
             ...createEmptyChangeMap<PaintData>(),
             updated: [
@@ -372,12 +354,10 @@ describe('useChatSevenTvCallbacks', () => {
     });
 
     test('records a Sentry count metric for badge updates', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onCosmeticUpdate(
+        result.onCosmeticUpdate(
           createBadgeCosmeticUpdateData({
             ...createEmptyChangeMap<BadgeData>(),
             updated: [
@@ -436,12 +416,10 @@ describe('useChatSevenTvCallbacks', () => {
 
   describe('onCosmeticDelete', () => {
     test('calls removeBadge and removePaint with cosmeticId', () => {
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onCosmeticDelete({
+        result.onCosmeticDelete({
           cosmeticId: 'cosmetic-123',
         });
       });
@@ -458,12 +436,10 @@ describe('useChatSevenTvCallbacks', () => {
         paintId: 'paint-1',
         badgeId: 'badge-1',
       });
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onEntitlementUpdate(updateData);
+        result.onEntitlementUpdate(updateData);
       });
 
       expect(applyEntitlementUpdateEvent).toHaveBeenCalledWith(updateData);
@@ -473,12 +449,10 @@ describe('useChatSevenTvCallbacks', () => {
   describe('onEntitlementDelete', () => {
     test('delegates entitlement deletes to the bridge', () => {
       const deleteData = createEntitlementDeleteData({ ttvUserId: 'ttv-1' });
-      const { result } = renderHook(() =>
-        useChatSevenTvCallbacks(defaultProps),
-      );
+      const result = createSevenTvCallbacks(defaultProps);
 
       act(() => {
-        result.current.onEntitlementDelete(deleteData);
+        result.onEntitlementDelete(deleteData);
       });
 
       expect(applyEntitlementDeleteEvent).toHaveBeenCalledWith(deleteData);
